+++ Monofoxe/Monofoxe.Engine/Vector3Extensions.cs
@@ -50,5 +50,8 @@ namespace Monofoxe.Engine
 			return v;
 		}
 
+
+		public static Vector3 Projection(this Vector3 v, Vector3 other) =>
+			Vector3.Dot(v, other) / Vector3.Dot(v, v) * v;
 	}
 }